requirements/integration.txt
# These dependencies are installed using:
#
#   pip install --upgrade
#
# This file should only declare dependencies related to code
# formatting, linting, testing and packaging.
#
# No dependencies required for production should be listed here.

# http://www.pydocstyle.org/en/4.0.0/release_notes.html
# Support for Python 2.x and PyPy has been dropped (#340).
pydocstyle<4


flake8
flake8-bugbear
flake8-docstrings
flake8-builtins
flake8-comprehensions
flake8-junit-report
pylint-ignore>=2020.1013
mypy
# pylint doesn't support isort>=5 for now
#   https://github.com/PyCQA/pylint/issues/3722
isort<5

# http://doc.pytest.org/en/latest/py27-py34-deprecation.html
# The pytest 4.6 series will be the last to support Python 2.7
# and 3.4, and is scheduled to be released by mid-2019.
# pytest 5.0 and onwards will support only Python 3.5+.
pytest<5.0
pytest-cov
# https://github.com/pytest-dev/pytest-html/blob/master/CHANGES.rst
# pytest-html 2.0+ doesn't support python2.7
pytest-html<2.0

readme_renderer[md]
twine

md-toc
straitjacket>=v202008.1016
bumpver
lib3to6

# needed for mypy coverage report
lxml

# optional needed for rendering png
Pillow
